+++ tests/unit/test_card_class.gd
@@ -64,17 +64,17 @@ func test_init_card_name():
 	assert_eq("Test Card 1",test1.card_name,
 			'card_name variable is set correctly')
 	assert_string_contains(test1.name, "Test Card 1")
-	assert_eq("Test Card 1",test1.get_node("Control/Front/CardText/Name").text,
+	assert_eq("Test Card 1",test1.get_node(Card.CARD_LABELS["Name"]).text,
 			'Name Label text is set correctly')
 	assert_eq("Test Card 2",test2.card_name,
 			'card_name variable is set correctly')
 	assert_string_contains(test2.name, "Test Card 2")
-	assert_eq("Test Card 2",test2.get_node("Control/Front/CardText/Name").text,
+	assert_eq("Test Card 2",test2.get_node(Card.CARD_LABELS["Name"]).text,
 			'Name Label text is set correctly')
 	assert_eq("Test Card 3",test3.card_name,
 			'card_name variable is set correctly')
 	assert_string_contains(test3.name, "Test Card 3")
-	assert_eq("Test Card 3",test3.get_node("Control/Front/CardText/Name").text,
+	assert_eq("Test Card 3",test3.get_node(Card.CARD_LABELS["Name"]).text,
 			'Name Label text is set correctly')
 
 func test_card_name_setget():
@@ -82,21 +82,21 @@ func test_card_name_setget():
 	assert_eq("Testing Name Change 1",card.card_name,
 			'card_name variable is set correctly')
 	assert_string_contains(card.name, "Testing Name Change 1")
-	assert_eq("Testing Name Change 1",card.get_node("Control/Front/CardText/Name").text,
+	assert_eq("Testing Name Change 1",card.get_node(Card.CARD_LABELS["Name"]).text,
 			'Name Label text is set correctly')
 	card.card_name = "Testing Name Change 2"
 	assert_eq("Testing Name Change 2",card.card_name,
 			'card_name variable is set correctly')
 	assert_string_contains(card.name, "Testing Name Change 2")
-	assert_eq("Testing Name Change 2",card.get_node("Control/Front/CardText/Name").text,
+	assert_eq("Testing Name Change 2",card.get_node(Card.CARD_LABELS["Name"]).text,
 			'Name Label text is set correctly')
 
 func test_CardDefinition_properties():
 	pending("Array property should use the separator")
 
 func test_font_size():
-	var text_node = card.get_node("Control/Front/CardText/Abilities")
-	var labels_rect = card.get_node("Control/Front/CardText").rect_size
+	var text_node = card.get_node(Card.CARD_LABELS["Abilities"])
+	var labels_rect = card._card_text.rect_size
 	var lorem = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed vitae laoreet nunc. Etiam vitae tempus ligula. Vestibulum pellentesque mauris vel ultricies pharetra. Curabitur iaculis dolor vitae leo aliquet viverra sit amet vitae eros. Nulla eros turpis, mollis non elit eget, iaculis porta magna. Vestibulum ante ipsum primis in faucibus orci luctus et ultrices posuere cubilia curae; Nulla sem dui, consectetur ut nulla ut, vestibulum venenatis dui. Fusce sollicitudin bibendum quam, at scelerisque risus fringilla eu. Curabitur elementum sem sed nisi malesuada molestie. Nulla finibus, eros quis volutpat vehicula, dui mauris dictum metus, quis pellentesque sem quam eget mi. Vivamus convallis massa non ex laoreet pharetra. Proin sed leo at dui varius dapibus sit amet ac purus."
 	card._set_label_text(text_node, lorem)
-	assert_eq(card.get_node("Control/Front/CardText").rect_size,labels_rect)
+	assert_eq(card._card_text.rect_size,labels_rect)